Add global requireADLogin option
Adds a global `requireADLogin` option to the ADAL config. You can
specify when configuring the provider:

    adalAuthenticationServiceProvider.init({
        tenant: '<tenant name>',
        clientId: '<client ID>',
        requireADLogin: true
    }, $httpProvider);

The global value for `requireADLogin` will be used whenever there is
**no** `requireADLogin` property on the route (or when it is `null` or
`undefined`). The route's `requireADLogin`  will always override the
global one so setting the route's `requireADLogin` property to `false`
will override the global `true`, resulting in AD login *not* being
required for this route.

This will make things easier when a lot of routes, if not all routes,
require login.