readme: add setup step addressing necessary config for HTML5Mode
When HTML5Mode is configured, if $locationProvider hashPrefix is
not set, the login sequence goes into an infinite loop. This
happens because, in HTML5Mode, the $locationProvider rewrites all
URLs to remove the hash+hashPrefix and AAD uses the hash to pass
the token via the redirect to the authenticating app. The logic
in ADAL.js fails to find the token as the hash is removed when
no hashPrefix is defined causing the token parse to fail and
ultimately the client-side authentication.