Added hash parameter to handleWindowCallback.
To allow for scenarios where the login flow takes place in a window
different than the main one, the handleWindowCallback function cannot
assume that the hash always comes from window.location or that
window.location should be replaced by a cached location after the
callback has been handled.

A hash parameter is introduced and will be used by the function if
present - otherwise it falls back to window.location.hash as before.
Also, window.location.hash and window.location are only overwritten
by the function when the hash in question was read from
window.location.