js: Add optional "version" parameter to config
AAD B2C requires "v2.0" in the authorization endpoint. There has not
been a way to make ADAL JS generate a login URL like
".../oauth2/v2.0/authorize...".

This commit adds an optional parameter "version" to the config object.
If it is set, it is injected in the login URL. If it is not set,
the behavior does not change to previous versions.

This is not full support for AAD B2C. However, it would be very
helpful until full support for AAD B2C comes for ADAL JS.