Fix typos, variable refs, and dead comment. fixes #309
There are a few typos in the comments, some variables that are referred to
that are not matched in casing, and an invalid comment that no longer
applies that make reading the code a bit confusing.

The comment about $injector was removed because it appears it no longer
applies. $injector is not used anymore in the service. There is no
circular dependancy issue.