full multi-tenant support
The existing token aquisition does not separate tokens, for the same
resource from different tenants, in the storage or the internal state.

The goal for these changes was to leave existing usage unchanged in
behavior and functionality.

To achieve that a new bool was added to the config
config.multiTenant
This flag defaults to false.

When that flag is true, a new public method can be used to
simultaneously request resource tokens from as many tenants as wanted.
acquireTokenForTenant(resource, tenantId, callback)

There is also a new method available to register callbacks
registerCallbackWithTenant(expectedState, resource, tenantId, callback)

Credit to Rajesh Babu Srikakollu for these changes