Only redirect to original url in case of login req
Redirection to the original page that was saved to storage before initiating the login should only happen when handling a login request. If the target page performs API calls, and the redirect is also performed when handling token renewals, this results in decreased performance due to the API calls being fired multiple times.