strip trailing # from uri when removing the hash
In a non-angular site, `handleWindowCallback` only removes `window.location.hash`, but not the trailing `#`.   I see in the dev branch, there has been a change made which seems to work when `navigateToLoginRequestUrl` is enabled, but introduces an unnecessary full page load. This is less than ideal for a single page app. Instead, if the browser supports history, just use replaceState() to avoid the full page reload.

Instead, it would be ideal to to use `window.history.replaceState()` if it is available.