fix: forwarding request without token to endpoint
There is situation when application wants to work with endpoint which offer some actions
which require authentication/authorization and some actions which don't.
It's good to forward requests to this type of endpoint because
actions will decide to accept request or not. Whitout this change, user
must be logged in if he/she want to send request to action of endpoint
which doesn't require authentication/authorization.