Reworking window assumptions made to broaden use
There were several points in handleWindowCallback() that made
assumptions on how adal was being used within the context of an
application. For instances such as within the context of a VSTS
Extension, operations like window.parent._adalInstance will be blocked
and result in CORS errors due to iframe behaviors. With the slight
reworking of the logic, adal can be used in instances such as previously
mentioned, to handle authentication for a VSTS Extension.

It's unclear whether these changes will suffice for a pull request, and
I'm hoping members of the Adal team can review and provide feedback, but
operations such as at line 1327 and 1337 probably need to be rethinked.