Make sure _loadFrameTimeout clean up _activeRenewals when facing timeout
Why:
We notice a bug with ADAL that sometimes, after browsers being sleep/idle for a while, ADAL gets into a weird state that "adal.token.renew.status" is "canceled", but new acquireToken calls are only registered in callbacks and not triggered.

How:
We update the _loadFrameTimeout function to make sure it will clean up the _activeRenewals for a given resource when timeout happens.